FMOD changes
added in music and ambience however I cant seem to change the switch values for my timeline of music. So the game music is set on the main menu song. Its also unbelievably loud for how loud I set it in fmod. Its legit -30 db and still seems loud af.